python_api/requirements.txt
pyserial>=3.5
cobs>=1.2.0
# Optional fast CRC backend:
# libscrc
# crcmod
